% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile.R
\name{datafile}
\alias{datafile}
\alias{dfn}
\title{return path of a file in the data directory}
\usage{
datafile(fname)
}
\arguments{
\item{fname}{filename (without the path)}
}
\value{
full path of file to data directory
}
\description{
Returns the full here() path from a file in the data directory
}
