% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafileInit.R
\name{datafileInit}
\alias{datafileInit}
\title{Initialise datafile}
\usage{
datafileInit(basedir = NA, subdir = NA, datadir = "data")
}
\arguments{
\item{basedir}{Path of the base directory}

\item{subdir}{character of vector of character with the names of
the subdirectory}

\item{datadir}{name of the data directory, default is 'data'}
}
\value{
full path of file to data directory
}
\description{
Data is stored in a subdirectory under the current workdir, or base
directory. This function sets the base directory, checks if it
exists and, if needed, creates the data directory
}
