% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile.R
\name{datafile}
\alias{datafile}
\alias{dfn}
\title{Expand a filename to a full path.}
\usage{
datafile(fname)
}
\arguments{
\item{fname}{filename}
}
\value{
full path of file to data directory

This function assumes that the data directory is correctly
initialized using datafileInit(). It will return a full path to the
file in the data directory.
}
\description{
This function expands a filename to a full file path
}
